+++ model.go
@@ -88,3 +88,12 @@ type mqttPublishing struct {
 func (m mqttPublishing) HashCode() string {
 	return m.Msg.MessageId
 }
+
+type RabbitMQEnvs struct {
+	Host     string `env:"RABBITMQ_HOST"`
+	Port     uint   `env:"RABBITMQ_PORT"`
+	Username string `env:"RABBITMQ_USERNAME"`
+	Password string `env:"RABBITMQ_PASSWORD"`
+	Vhost    string `env:"RABBITMQ_VHOST"`
+	UseTLS   bool   `env:"RABBITMQ_USE_TLS"`
+}